Consolidate legacy governance commands
